Fix loop variable types in linear sieve function

+++ sieve.cpp
@@ -4,10 +4,10 @@ vector<int>p;//have all prime numbers less than N.
 void linear_sieve()
 {
     isPrime[0]=isPrime[1]=0;
-    for(ll i=2;i<N;i++)
+    for(int i=2;i<N;i++)
     {
         if(isPrime[i]==1) p.push_back(i);
-        for(ll o=0;o<p.size();o++)
+        for(int o=0;o<p.size();o++)
         {
             if(i*p[o]>=N) break;
             isPrime[i*p[o]]=0;
